miconsu/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coverage==6.4.1
coveralls==3.3.1
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
zipp==3.8.0
